scraper/requirements.txt
scrapy>=2.11
trafilatura>=1.9
bleach>=6.1
supabase>=2.4
python-dotenv>=1.0
Pillow>=10.0
numpy>=1.26
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
